adminbot/requirements.txt
python-telegram-bot==12.8
pylint==2.6